backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI Integrations
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1
httpx==0.25.2

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
python-json-logger==2.0.7

# Audio Processing
soundfile==0.12.1
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2